rename and update tests
