+++ cmd/jimmsrv/main.go
@@ -204,7 +204,7 @@ func start(ctx context.Context, s *service.Service) error {
 			return jimmsvc.OpenFGACleanup(ctx, time.NewTicker(6*time.Hour).C)
 		})
 		s.Go(func() error {
-			return jimmsvc.WatchModelsDying(ctx, time.NewTicker(time.Minute).C)
+			return jimmsvc.CleanupModelsDying(ctx, time.NewTicker(time.Minute).C)
 		})
 	}
